% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(x,y,n=100,color="gold")
}
\arguments{
\item{x}{variable from dataset}

\item{y}{variable from dataset}

\item{n}{number of bootstap replicates}

\item{plot}{defines whether a is wanted}

\item{color}{defines the color of the confidence interval on plot}
}
\value{
List of true correlation, estimated correlation, standard error, bias and confidence interval
}
\description{
This function estimates the correlation using bootstrap.
}
\examples{
bootstrap(cars$speed,cars$dist)
}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}
\keyword{bootstrap}

