% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstraptask1.R
\name{Bootstraptask1}
\alias{Bootstraptask1}
\title{Bootstrap}
\usage{
Bootstraptask1(n,x,y)
}
\arguments{
\item{n}{is the number of bootstrap estimations}

\item{x}{is a vector of a given data set that you want to find a relation with another set}

\item{y}{is another vector of that you want to see if it has a relation to the first data set}
}
\value{
The function returns the estimated correlation, standard error, bias and 95 % confidence interval
}
\examples{
 Bootstraptask1(500,Data$Sport,Data$Grades) reads two vectors from the Data with 500 estimations
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Bootstrap}
\keyword{confidence_interval}
\keyword{correlation}

