% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.needle.R
\name{buffon.needle}
\alias{buffon.needle}
\title{buffons needle}
\usage{
buffon.needle(n, l, d, dig)
}
\arguments{
\item{n}{number of needles.}

\item{l}{length of needles}

\item{d}{distance between parallel lines}

\item{dig}{decimal places returned}
}
\description{
Find pi
}
\author{
Maria
}

