% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Buffontask2.R
\name{Buffontask2}
\alias{Buffontask2}
\title{Buffon needle experiment}
\usage{
Buffontask2(N,l,d)
}
\arguments{
\item{N}{is the amount of needles thrown}

\item{l}{is the length of the needles}

\item{d}{is the distance between the parallel lines}
}
\value{
the estimated pi values with a certain amount of thrown needles
}
\description{
Buffon's needle experiment a way of estimating pi by throwing needles on a floor with parallel lines and count the number of times the needle hits the lines.
}
\examples{
add the example text
Buffontask2(100,1,2)
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Buffon's_needle}
\keyword{pi}

