% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.GOF.R
\name{chi.GOF}
\alias{chi.GOF}
\title{Chi Probability}
\usage{
chi.GOF(x,p = 1/length(x))
}
\arguments{
\item{x}{categorical variable}

\item{p}{expected probability}
}
\value{
Goodness of fit value
}
\description{
This function computes the chi squared goodness of fit.
}
\examples{
chi.GOF(cars$speed)
}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}
\keyword{chi}
\keyword{fit}
\keyword{goodness}
\keyword{of}

