% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.probability.R
\name{chi.probability}
\alias{chi.probability}
\title{chi-square}
\usage{
chi.probability(x, df = 1, n)
}
\arguments{
\item{x}{numberical input.}

\item{df}{degress of freedom}

\item{n}{an integer}
}
\description{
Find chi square probability
}
\author{
Maria
}

