% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiprobability.R
\name{chiprobability}
\alias{chiprobability}
\title{Chi probability}
\usage{
chiprobability(x, df, n)
}
\arguments{
\item{x}{is the parameter that you want the chi square distribution to be larger than}

\item{df}{is degrees of dreedom in the Chi square distribution}

\item{n}{is the amount of generated data from the chi square distribution}
}
\value{
The function returns the probability that the chi square distribution with a given degree of freedom is larger than a given x
}
\examples{
chiprobability(1.5, 2, 10000) which return that there is approximately 47 percent chance that the chi square distribution is larger than 1.5 with 2 degrees of freedom
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Carlo}
\keyword{Chi-squared}
\keyword{and}
\keyword{distribution}
\keyword{estimation}
\keyword{monte}
\keyword{probability}

