% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{Density}
\alias{Density}
\title{Density estimator}
\usage{
Density(x,d,h,method = "naive")
}
\arguments{
\item{x}{numeric variable}

\item{d}{point you are evaluating to obtain its density}

\item{h}{bandwidth}

\item{method}{method for estimating the density. Is either "naive" or "kernel". Set default as "naive".}
}
\value{
Return, depending on input, either the density in a point or the bandwidth and
a table of densities estimated at minimum, 1st quartile, median, mean, 3rd quartile, maximum.
}
\description{
This function estimates the density of the given data.
}
\examples{
Density(cars$speed,3,0.33,method = "kernel")

}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}

