% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estdens.R
\name{estdens}
\alias{estdens}
\title{Density estimation}
\usage{
estdens(x, d, h, method)
}
\arguments{
\item{x}{is the data set you want to obtain probabilities from}

\item{d}{is the point in which you want to approximate the probability}

\item{h}{is the bandwidth which is the width of the categories that you want to the data into.}

\item{method}{is either naive or kernel depending on which method you want to use.}
}
\value{
a probability of a given point d if a d value is given. If the d value is not defined it returns the probability of the quantiles and the mean
}
\examples{
estdens(rgamma(10000,1), 1, 1, "naive") which returns the probability of 1 with a bandwidth of 1 and with a random gamma distribution with mean 1.
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Density_estimation}
\keyword{Gaussian_Kernel}
\keyword{Sturges_naive}

