% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goodnessoffittask3.R
\name{Goodnessoffittask3}
\alias{Goodnessoffittask3}
\title{Goodness of fit}
\usage{
Goodnessoffittask3(x, p)
}
\arguments{
\item{x}{is a vector of a given data set that you want to test for a given probability}

\item{p}{is the given probability of a distribution for each data of x}
}
\value{
The function returns the chi squared test value T
}
\examples{
reads two vectors from the nData which is the data of the soccer cards and then returns the Goodness of fit test value
Goodnessoffittask3(c(1,2,3),c(0.33,0.33,0.33))
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Chi-squared}
\keyword{Goodness_of_fit}

