% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.model.R
\name{linear.model}
\alias{linear.model}
\title{Linear Model}
\usage{
linear.model(formula)
}
\arguments{
\item{formula}{formula of the form dependent~predictor1+predictor2+...+predictorN}
}
\value{
A list with residuals, the estimate, standard error, t-value, p-value, R squared, Adjusted R squared and an F-statistic
}
\description{
This function creates a linear model.
}
\examples{
linear.model(cars$speed,cars$dist)
}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}

