% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Markovchain.R
\name{Markovchain}
\alias{Markovchain}
\title{Markov chain}
\usage{
Markovchain(p, k, n)
}
\arguments{
\item{p}{is the transition probability matrix where index [1,1] is the probability that state 1 returns to state 1 again, [1,2] is the probablity that state 1 goes to state 2 and so on}

\item{k}{is an integer indicating the initial state}

\item{n}{is the number of simulated steps}
}
\value{
a table with the stationary probability of each state
}
\examples{
Markovchain(A,3,10000) and returns the probability for each state for 10000 steps.
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{discrete_Markov_chain}
\keyword{transition_probability_matrix}

