% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrixregression.R
\name{Matrixregression}
\alias{Matrixregression}
\title{Multiple linear regression}
\usage{
Matrixregression(dependent ~ predictor1 + predictor2 + ...)
}
\arguments{
\item{formula}{is the dependent variable as a function of the predictors.}

\item{Dependent}{is the variable you want to check if it fits a linear model of the predictors}

\item{Predictors}{are the variables that you want to see if there is a linear model between them and the dependent variable}
}
\value{
2 tables, the first with coefficients for a linear model fitting the data points for each of those coefficents there is a standard error, t-value and p-value. The second table returns the residual standard error, R^2, R^2adj, F-statistic, degrees of freedom, p-value.
}
\examples{
Matrixregression(cars$speed~cars$dist) which returns a linear function of speed dependent on the break distance.
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Matrix_regression}
\keyword{Multiple_linear_regression}
\keyword{Residual}
\keyword{formula}

