% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdens.R
\name{plotdens}
\alias{plotdens}
\title{Density plot}
\usage{
plotdens(x, n, method, from, to)
}
\arguments{
\item{x}{is the data set you want to make a density plot of}

\item{n}{is the number of points used for plotting the density function.}

\item{method}{is either naive or kernel depending on which method you want to use.}

\item{from}{is the start point of your plot.}

\item{to}{is the end point of your plot.}
}
\value{
a density plot of a data set with a given method
}
\examples{
plotdens(rgamma(10000,1), 500, "naive", 0, 6) returns a density plot of a gamma distribution with mean 1 and the naive method is used in the inverval between 0 and 6.
}
\author{
Sebastian Gregersen and Jeppe Korsgaard Kristensen
}
\keyword{Density_plot}
\keyword{Sturges_naive}
\keyword{gaussian_kernel}

