% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffon.needle.R
\name{buffon.needle}
\alias{buffon.needle}
\title{Buffon's Needle}
\usage{
buffon.needle(N = 1000, l = 1, d = 1)
}
\arguments{
\item{N}{the number of needles thrown}

\item{l}{the length of the needle}

\item{d}{the distance between the stripes on the floor}
}
\value{
Estimated pi and the probability of hit
}
\description{
The function estimates pi by simulating a throw with needles.
}
\examples{
buffon.needle(3000,l = 2, d = 4)
}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}
\keyword{buffon}
\keyword{needle}
\keyword{pi}

