% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.R
\name{MarkovChain}
\alias{MarkovChain}
\title{Markov Chain}
\usage{
MarkovChain(p,k,n)
}
\arguments{
\item{p}{transition probability matrix}

\item{k}{integer indicating the initial state of the chain}

\item{n}{number of simulated steps}
}
\value{
Stationary probabilities for each state
}
\description{
This function simulates the stationary probabilities from a transition probability matrix.
}
\examples{
MarkovChain(matrix(c(.5,.5,.5,.5),ncol=2,byrow=TRUE),1,1000)
}
\author{
Nina Louise Pedersen \cr
Institute of Mathematics and Computer Science (IMADA) \cr
University of Sourthern Denmark, (SDU) \cr
\email{npede14@student.sdu.dk} \cr
}
\keyword{chain}
\keyword{markov}
\keyword{stationary}

